webpack: Move webpack configuration to web.
