+++ web/babel.config.js
