+++ web/webpack.assets.json
@@ -1,89 +1,86 @@
 {
     "activity": [
-        "./web/src/bundles/common",
+        "./src/bundles/common",
         "sorttable",
-        "./web/src/analytics/activity",
-        "./web/styles/portico/activity.css"
+        "./src/analytics/activity",
+        "./styles/portico/activity.css"
     ],
     "billing": [
-        "./web/src/bundles/portico",
-        "./web/src/portico/landing-page",
-        "./web/styles/portico/landing_page.css",
-        "./web/src/billing/helpers",
-        "./web/src/billing/billing",
-        "./web/src/templates",
-        "./web/src/loading",
-        "./web/styles/portico/billing.css"
+        "./src/bundles/portico",
+        "./src/portico/landing-page",
+        "./styles/portico/landing_page.css",
+        "./src/billing/helpers",
+        "./src/billing/billing",
+        "./src/templates",
+        "./src/loading",
+        "./styles/portico/billing.css"
     ],
     "upgrade": [
-        "./web/src/bundles/portico",
-        "./web/src/portico/tippyjs",
-        "./web/src/portico/landing-page",
-        "./web/styles/portico/landing_page.css",
-        "./web/src/billing/helpers",
-        "./web/src/billing/upgrade",
-        "./web/src/templates",
-        "./web/src/loading",
-        "./web/styles/portico/billing.css"
+        "./src/bundles/portico",
+        "./src/portico/tippyjs",
+        "./src/portico/landing-page",
+        "./styles/portico/landing_page.css",
+        "./src/billing/helpers",
+        "./src/billing/upgrade",
+        "./src/templates",
+        "./src/loading",
+        "./styles/portico/billing.css"
     ],
     "billing-event-status": [
-        "./web/src/bundles/portico",
-        "./web/styles/portico/landing_page.css",
-        "./web/src/billing/event_status.js",
-        "./web/src/billing/helpers",
-        "./web/styles/portico/billing.css"
+        "./src/bundles/portico",
+        "./styles/portico/landing_page.css",
+        "./src/billing/event_status.js",
+        "./src/billing/helpers",
+        "./styles/portico/billing.css"
     ],
-    "portico": ["./web/src/bundles/portico"],
-    "error-styles": [
-        "./web/third/bootstrap/css/bootstrap.css",
-        "./web/styles/portico/portico_styles.css"
-    ],
-    "common": ["./web/src/bundles/common"],
+    "portico": ["./src/bundles/portico"],
+    "error-styles": ["./third/bootstrap/css/bootstrap.css", "./styles/portico/portico_styles.css"],
+    "common": ["./src/bundles/common"],
     "help": [
-        "./web/src/bundles/portico",
+        "./src/bundles/portico",
         "simplebar/dist/simplebar.css",
         "simplebar",
-        "./web/src/portico/help",
-        "./web/src/portico/tabbed-instructions"
+        "./src/portico/help",
+        "./src/portico/tabbed-instructions"
     ],
     "landing-page": [
-        "./web/src/bundles/portico",
-        "./web/src/portico/landing-page",
-        "./web/styles/portico/landing_page.css"
+        "./src/bundles/portico",
+        "./src/portico/landing-page",
+        "./styles/portico/landing_page.css"
     ],
     "integrations": [
-        "./web/src/bundles/portico",
-        "./web/src/portico/integrations",
-        "./web/styles/portico/landing_page.css",
-        "./web/styles/portico/integrations.css"
+        "./src/bundles/portico",
+        "./src/portico/integrations",
+        "./styles/portico/landing_page.css",
+        "./styles/portico/integrations.css"
     ],
     "communities": [
-        "./web/src/bundles/portico",
-        "./web/src/portico/communities",
-        "./web/styles/portico/landing_page.css",
-        "./web/styles/portico/integrations.css"
+        "./src/bundles/portico",
+        "./src/portico/communities",
+        "./styles/portico/landing_page.css",
+        "./styles/portico/integrations.css"
     ],
-    "signup": ["./web/src/bundles/portico", "jquery-validation", "./web/src/portico/signup"],
-    "register": ["./web/src/bundles/portico", "jquery-validation", "./web/src/portico/signup"],
+    "signup": ["./src/bundles/portico", "jquery-validation", "./src/portico/signup"],
+    "register": ["./src/bundles/portico", "jquery-validation", "./src/portico/signup"],
     "confirm-preregistrationuser": [
-        "./web/src/bundles/common",
-        "./web/src/portico/confirm-preregistrationuser"
+        "./src/bundles/common",
+        "./src/portico/confirm-preregistrationuser"
     ],
     "support": [
-        "./web/src/bundles/common",
+        "./src/bundles/common",
         "sorttable",
-        "./web/src/analytics/activity",
-        "./web/styles/portico/activity.css",
-        "./web/src/analytics/support"
+        "./src/analytics/activity",
+        "./styles/portico/activity.css",
+        "./src/analytics/support"
     ],
-    "desktop-login": ["./web/src/bundles/portico", "./web/src/portico/desktop-login"],
-    "desktop-redirect": ["./web/src/bundles/portico", "./web/src/portico/desktop-redirect"],
+    "desktop-login": ["./src/bundles/portico", "./src/portico/desktop-login"],
+    "desktop-redirect": ["./src/bundles/portico", "./src/portico/desktop-redirect"],
     "stats": [
-        "./web/src/bundles/portico",
-        "./web/styles/portico/stats.css",
-        "./web/src/stats/stats",
+        "./src/bundles/portico",
+        "./styles/portico/stats.css",
+        "./src/stats/stats",
         "tippy.js/dist/tippy.css"
     ],
-    "app": ["./web/src/bundles/app"],
-    "digest": ["./web/src/bundles/portico"]
+    "app": ["./src/bundles/app"],
+    "digest": ["./src/bundles/portico"]
 }
