+++ web/webpack.dev-assets.json
@@ -1,17 +1,17 @@
 {
-    "dev-login": ["./web/src/bundles/portico", "./web/src/portico/dev-login"],
+    "dev-login": ["./src/bundles/portico", "./src/portico/dev-login"],
     "dev-integrations-panel": [
-        "./web/src/bundles/portico",
-        "./web/src/portico/integrations_dev_panel",
-        "./web/styles/portico/integrations_dev_panel.css",
-        "./web/src/reload_state",
-        "./web/src/channel"
+        "./src/bundles/portico",
+        "./src/portico/integrations_dev_panel",
+        "./styles/portico/integrations_dev_panel.css",
+        "./src/reload_state",
+        "./src/channel"
     ],
     "dev-email-log": [
-        "./web/src/bundles/common",
-        "./web/src/portico/email_log",
-        "./web/styles/portico/email_log.css",
-        "./web/src/reload_state",
-        "./web/src/channel"
+        "./src/bundles/common",
+        "./src/portico/email_log",
+        "./styles/portico/email_log.css",
+        "./src/reload_state",
+        "./src/channel"
     ]
 }
